hitbox: split up extensions
